Use indenting reporter to debug Cocoapods tests
